+++ programs/amm/src/endpoints/farming.rs
@@ -1,3 +1,3 @@
-pub mod create_lp_vault;
+pub mod create_farm;
 
-pub use create_lp_vault::*;
+pub use create_farm::*;