+++ programs/amm/src/endpoints/farming/create_farm.rs
@@ -1,0 +1,10 @@
+use crate::prelude::*;
+
+#[derive(Accounts)]
+pub struct CreateFarm<'info> {
+    pub admin: Signer<'info>,
+}
+
+pub fn handle(_ctx: Context<CreateFarm>) -> Result<()> {
+    Ok(())
+}