+++ programs/amm/src/models.rs
@@ -1,6 +1,8 @@
 pub mod farm;
+pub mod farmer;
 
 pub use farm::*;
+pub use farmer::*;
 
 use crate::prelude::*;
 
@@ -16,4 +18,32 @@ use crate::prelude::*;
 )]
 pub struct TokenAmount {
     pub amount: u64,
+}
+
+#[derive(
+    AnchorDeserialize,
+    AnchorSerialize,
+    Clone,
+    Copy,
+    Debug,
+    Default,
+    Eq,
+    PartialEq,
+)]
+pub struct Slot {
+    pub slot: u64,
+}
+
+#[derive(
+    AnchorDeserialize,
+    AnchorSerialize,
+    Clone,
+    Copy,
+    Debug,
+    Default,
+    Eq,
+    PartialEq,
+)]
+pub struct MintHash {
+    pub hash: u64,
 }