+++ programs/amm/CHANGELOG.md
@@ -1,12 +1,21 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-05-20
+
+### Added
+
+- Model for `Farm` according to the design doc.
+- Model for `Farmer` according to the design doc.
 
 ## [0.1.0] - 2022-05-12
+
 ### Added
+
 - Skeleton for models, endpoints and other modules. The skeleton is analogous
-    to our other programs.
+  to our other programs.
 - A model for `Farm` which enables farming of a specific harvest mint.